+++ src/articles/2027-01-01-bug-bodycam.mdx
@@ -27,7 +27,7 @@ Article text here etc etc Article text here etc etc Article text here etc etc Ar
 Article text here etc etc Article text here etc etc Article text here etc etc Article text here etc etc
 Article text here etc etc Article text here etc etc Article text here etc etc Article text here etc etc
 
-<VideoExpander src="/bug-video/website_cut.mp4" />
+<BugVideoExpander />
 
 Article text here etc etc Article text here etc etc Article text here etc etc Article text here etc etc
 Article text here etc etc Article text here etc etc Article text here etc etc Article text here etc etc
