+++ src/components/layout.js
@@ -76,8 +76,8 @@ import Wikimage from '../visuals/wikiedits-visuals/wikimage';
 import DatawrapperChart from './dataWrapper';
 import ASUCFunding from '../visuals/club-finances-visuals/asucFunding';
 import SexyMap from '../visuals/too-sexy-visuals/map';
-import { BugVideoScroller } from '../visuals/bug-video-visuals/bug-video-scroller';
-import { VideoExpander } from '../visuals/bug-video-visuals/video-expander';
+import { BugVideoScroller } from '../visuals/bug-bodycam-visuals/bug-video-scroller';
+import { BugVideoExpander } from '../visuals/bug-bodycam-visuals/bug-video-expander';
 /*
 To avoid using exact paths in MDX files, import your components here
 and add them to the shortcodes list to be globally accessible.ß
@@ -167,7 +167,7 @@ const shortcodes = {
   ASUCFunding,
   SexyMap,
   BugVideoScroller,
-  VideoExpander,
+  BugVideoExpander,
 };
 
 export default function Layout({ children }) {
