+++ src/visuals/bug-bodycam-visuals/bug-video-expander.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { VideoExpander } from './components/video-expander';
+
+export const BugVideoExpander = () => (
+  <VideoExpander src="/bug-video/website_cut.mp4" />
+);
